Several adjustments.
Added an Options GUI for handling some config settings.
Added a config option to disable/enable the timer.
Playing and pausing of the tickerTimeline is now handled via listening to its statusProperty as well as setting the buttons text.
Refactored the loading of the ApplicationConfig into a separate method which re-initializes some values during runtime when the config changes.
Ticker provides a getter to see if the session is currently paused.