Added a "Stop" button.
In order to properly implement this function I had to refactor the way the session progress bar gets handled. Its progress is now bound to a progressProperty for easier handling throughout the code. Additionally MainController.initializeTicker now also resets the progress and timer. And lastly a bug in the tickerTimeline status property listener has been fixed where the timerTimeline would get resumed even if it was stopped (which should not happen).